tech debt and prop types cleanup
