+++ frontend/src/Components/ProfileNavigationMenu/ProfileNavigationMenu.jsx
@@ -1,43 +1,8 @@
+import { useState } from 'react';
 import { Box, Divider, makeStyles } from '@material-ui/core';
 import { NavigationTabBar } from '../Event/EventDetailsDrawerComponent';
-import { BookmarkRounded, ListRounded, NotesRounded, TimelineRounded } from '@material-ui/icons';
-import { useState } from 'react';
 import RecentActivitiesList from '../RecentActivitiesList/RecentActivitiesList';
-
-export const PROFILE_NAVIGATION_MENU_BAR = [
-  {
-    id: 1,
-    icon: <BookmarkRounded />,
-    displayName: 'Tags',
-    subtitle: '',
-    color: '',
-    onClick: () => {},
-  },
-  {
-    id: 2,
-    icon: <TimelineRounded />,
-    displayName: 'Recent Activities',
-    subtitle: 'Recently created events or volunteered events',
-    color: '',
-    onClick: () => {},
-  },
-  {
-    id: 3,
-    icon: <ListRounded />,
-    displayName: 'Inventories',
-    subtitle: '',
-    color: '',
-    onClick: () => {},
-  },
-  {
-    id: 4,
-    icon: <NotesRounded />,
-    displayName: 'Notes',
-    subtitle: '',
-    color: '',
-    onClick: () => {},
-  },
-];
+import { PROFILE_NAVIGATION_MENU_BAR } from './constants';
 
 const useStyles = makeStyles((theme) => ({
   spinnerContainer: {
